requirements-dev.txt
-r requirements.txt
ipython
django-extensions
django-debug-toolbar
